pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aifinreport"
version = "0.1.0"
description = "AI-powered financial news report generator using LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "mistralai>=1.9.0",
    "openai>=2.0.0",
    "httpx>=0.28.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.14.0",
    "trafilatura>=2.0.0",
    "scikit-learn>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
aifinreport = "aifinreport.cli.generate_report:main"

[tool.setuptools.packages.find]
where = ["src"]
